+++ app/src/main/kotlin/com/simplemobiletools/calendar/pro/activities/SplashActivity.kt
@@ -14,7 +14,7 @@ class SplashActivity : BaseSplashActivity() {
             }
             intent.extras?.containsKey(EVENT_ID) == true -> Intent(this, MainActivity::class.java).apply {
                 putExtra(EVENT_ID, intent.getLongExtra(EVENT_ID, 0L))
-                putExtra(EVENT_OCCURRENCE_TS, intent.getIntExtra(EVENT_OCCURRENCE_TS, 0))
+                putExtra(EVENT_OCCURRENCE_TS, intent.getLongExtra(EVENT_OCCURRENCE_TS, 0L))
                 startActivity(this)
             }
             else -> startActivity(Intent(this, MainActivity::class.java))
